Move OutputsCleaner to the history package
