+++ subprojects/execution/src/main/java/org/gradle/internal/execution/history/OutputsCleaner.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 the original author or authors.
+ * Copyright 2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.execution.impl;
+package org.gradle.internal.execution.history;
 
 import com.google.common.collect.Ordering;
 import org.gradle.internal.file.Deleter;
